fix and generalize uploader script
