+++ api-service/api-server/contrib_typings/google-cloud__spanner/index.d.ts
